+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - `plot_melt_curves_coupled.R` for single-oligo coupled melting curve plot.
+- Allowed for initial formamide and sodium concentration for melting temperature correction.
 
 ## [1.4.4]
 ### Changed